// src/models/Lead.ts
export type Lead = {
  leadId: number;
  leadFirstName: string;
  leadLastName: string;
  leadFullName: string;
  leadSuburb: string;
  leadZipCode: string;
  leadCategory: string;
  leadDescription: string;
  leadPhoneNumber: string;
  leadEmailAddress: string;
  leadPrice: number;
  leadCreatedAt: string;
  leadUpdatedAt: string;
  statusId: number;
  statusName: string;
}